% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{rn5}
\alias{rn5}
\title{RN5 - Annotation tables}
\format{\code{list} of four elements - \code{"regions"}, \code{"sites"}, \code{"controls"} and \code{"mappings"}.
        These elements are described below.
        \describe{
          \item{\code{"regions"}}{\code{list} of \code{NULL}s; the names of the elements correspond to the built-in
               region annotation tables. Once the default annotations are loaded, the attribute \code{"builtin"} is
               a \code{logical} vector storing, for each region annotation, whether it is the default (built-in) or
               custom.}
          \item{\code{"sites"}}{\code{list} of \code{NULL}s; the names of the elements correspond to the site
               annotation tables.}
          \item{\code{"mappings"}}{\code{list} of \code{NULL}s; the names of the elements correspond to the built-in
               region annotation tables.}
        }}
\description{
Scaffold of annotation tables for RN5. This structure is automatically loaded upon initialization of the annotation,
that is, by the first valid call to any of the following functions: \code{\link{rnb.get.assemblies}},
\code{\link{rnb.get.chromosomes}}, \code{\link{rnb.get.annotation}}, \code{\link{rnb.set.annotation}},
\code{\link{rnb.get.mapping}}, \code{\link{rnb.annotation.size}}. Adding an annotation amounts to attaching its
table(s) and mapping structures to this scaffold.
}
\author{
Yassen Assenov
}
\keyword{datasets}

